Refactored the SearchOrganisationView and templates to reflect the change in path.
