+++ aliss/templates/organisation/organisation-search.html
@@ -12,7 +12,7 @@
             <label for="postcode">Please enter keywords to find more information about organisations.</label>
             <div class="search-box">
                 <div class="search-holder sbs-inputs">
-                    <form action="{% url 'results' %}">
+                    <form action="{% url 'organisation_search_results' %}">
                         <input class="query" type="text" name="q" id="q" placeholder="e.g. Care Scotland"
                         <div class="search-button">
                             <button type="submit"><span>Search</span></button>
