+++ aliss/templates/organisation/search-results.html
@@ -12,7 +12,7 @@
         <h1>Organisations related to: <span class="postcode">{{ request.GET.q }}</span>
         </h1>
         <div class="buttons">
-          <a href="{% url 'organisations' %}" class="button primary">New organisation search</a>
+          <a href="{% url 'organisation_search' %}" class="button primary">New organisation search</a>
         </div>
       </div>
     </div>
